Refactor SearchController to service
